a working state of code
